TIP-842: Move Attribute Group Translation model
