+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -17,7 +17,7 @@ parameters:
     pim_catalog.entity.attribute.class:                    Akeneo\Pim\Structure\Component\Model\Attribute
     pim_catalog.entity.attribute_translation.class:        Pim\Bundle\CatalogBundle\Entity\AttributeTranslation
     pim_catalog.entity.attribute_group.class:              Akeneo\Pim\Structure\Component\Model\AttributeGroup
-    pim_catalog.entity.attribute_group_translation.class:  Pim\Bundle\CatalogBundle\Entity\AttributeGroupTranslation
+    pim_catalog.entity.attribute_group_translation.class:  Akeneo\Pim\Structure\Component\Model\AttributeGroupTranslation
     pim_catalog.entity.attribute_option.class:             Akeneo\Pim\Structure\Component\Model\AttributeOption
     pim_catalog.entity.attribute_option_value.class:       Pim\Bundle\CatalogBundle\Entity\AttributeOptionValue
     pim_catalog.entity.attribute_requirement.class:        Akeneo\Pim\Structure\Component\Model\AttributeRequirement
